Extract communication methods to `AbstractSendable`
